Move the simplifier tests out of libHalide
Simplify.cpp is too large, and the simplifier tests don't use any
internal functionality anyway.

The contents of simplify.cpp is copy-pasted from Simplify.cpp in master.
