+++ src/IREquality.cpp
@@ -201,7 +201,55 @@ IRComparer::CmpResult IRComparer::compare_types(Type a, Type b) {
     compare_scalar(a.code(), b.code());
     compare_scalar(a.bits(), b.bits());
     compare_scalar(a.lanes(), b.lanes());
-    compare_scalar((uintptr_t)a.handle_type, (uintptr_t)b.handle_type);
+
+    if (result != Equal || !a.is_handle()) return result;
+
+    const halide_handle_cplusplus_type *ha = a.handle_type;
+    const halide_handle_cplusplus_type *hb = b.handle_type;
+
+    if (ha == hb) {
+        // Early-out
+        return result;
+    }
+
+    if (ha == nullptr) {
+        result = LessThan;
+        return result;
+    }
+
+    if (hb == nullptr) {
+        result = GreaterThan;
+        return result;
+    }
+
+    compare_scalar(ha->reference_type, hb->reference_type);
+    compare_names(ha->inner_name.name, hb->inner_name.name);
+    compare_scalar(ha->inner_name.cpp_type_type, hb->inner_name.cpp_type_type);
+    compare_scalar(ha->namespaces.size(), hb->namespaces.size());
+    compare_scalar(ha->enclosing_types.size(), hb->enclosing_types.size());
+    compare_scalar(ha->cpp_type_modifiers.size(), hb->cpp_type_modifiers.size());
+
+    if (result != Equal) return result;
+
+    for (size_t i = 0; i < ha->namespaces.size(); i++) {
+        compare_names(ha->namespaces[i], hb->namespaces[i]);
+    }
+
+    if (result != Equal) return result;
+
+    for (size_t i = 0; i < ha->enclosing_types.size(); i++) {
+        compare_scalar(ha->enclosing_types[i].cpp_type_type,
+                       hb->enclosing_types[i].cpp_type_type);
+        compare_names(ha->enclosing_types[i].name,
+                      hb->enclosing_types[i].name);
+    }
+
+    if (result != Equal) return result;
+
+    for (size_t i = 0; i < ha->cpp_type_modifiers.size(); i++) {
+        compare_scalar(ha->cpp_type_modifiers[i],
+                       hb->cpp_type_modifiers[i]);
+    }
 
     return result;
 }
