+++ src/Simplify.h
@@ -95,9 +95,6 @@ template<> inline double div_imp<double>(double a, double b) {
     return a/b;
 }
 
-
-void simplify_test();
-
 }
 }
